Initial implementation of a Zookeeper-based ReferenceDatabase
Working towards a distributed Git setup, supports multiple different
master servers, typically in different geographical locations
("replicas"). Each replica has its own local copy of the repository
and references, and serializes updates to a single global version
counter node in Zookeeper. When a replica receives a ref update, it
first ensures its local copy is up to date with respect to the global
version, then atomically writes out a version update along with enough
information for every replica to come up to date.

Each replica watches the global version node and applies updates as
they come in, fetching data from the source of the update as
necessary. Eventually, we want to support quorum object writes, such
that a ref update cannot be committed to Zookeeper until the data is
written to a majority of replicas.

Change-Id: I115b8564a80bebe04b1f237cd462a195ef53e816