Add tests for zookeeper auth config parsing
We have now split parsing of the curator config from actually building
the curator object.

When calling builder.build() the Curator framework waits for it to
connect to a zookeeper instance, which we don't have in this instance.

In order to be able to test the parsed configuration without the need to
start a Zookeeper instance, the refactoring in ZookeeperConfig.java was
required.

Change-Id: I0695678348e64b7fc7390d4eb5bbcec19f290475