Honour secure configs during init step
During the init step, zookeeper-refdb might need to connect to the
Zookeeper ensemble to perform migration-related tasks, as it was
introduced in I7836b739bb.

When doing so however, the curator was being initialized without
honouring the configurations set in the zookeeper-refdb.secure.config,
as it was reading the zookeeper-refdb.config only.

Address this issue by reading from zookeeper-refdb.secure.config and
falling back to the zookeeper-refdb.config.

Change-Id: I9123a979e01f4811dbd71cc41445eb109cf6bfae